requirements.txt
alembic==1.16.1
Faker==37.3.0
greenlet==3.2.2
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dotenv==1.1.0
setuptools==80.8.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
